+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ### v0.7.0
 
+#### Add
+
 - Added encrypted database values for QA and production.
+
+#### Update
+
+- Updated the return object response for multiple address for the `validations/address` and `patron` endpoints.
 
 ### v0.6.0
 
@@ -12,7 +18,7 @@
 - Added a request flag for terms of condition acceptance input.
 - Added a request flag for the age gate input for "simplye" policy types.
 
-#### Updated
+#### Update
 
 - Updated endpoint `/patrons` to use `workAddress` and to use the logic for creating the right card for API V0.3.
 - Updated the address validations to include a work address as well as implementing validation through Service Objects.